Version 2.0.0 (#5)
* Improved Frontend and Camaras service

* Improved Frontend and Camaras service

* Fix Firewalls error - state:dead

* Fix Firewalls error - state:dead

* Listos servicios WAN y FW

* Add Login module (#2)

* Listo el modulo para crear, editar y eliminar usuarios

* Listo el modulo para crear, editar y eliminar usuarios

* Ready login module

* Ready login module

* Ready Login Module

* Fix button DCS

* Fix error edit users

* Changes Firewalls and Wan modules (#3)

* Guardando cambios en la rama fix-wan-fw

* Ready for deploy in production, FW, WAN, Backend, Frontend

* Ready for deploy in production, FW, WAN, Backend, Frontend

* Ready for deploy in production, FW, WAN, Backend, Frontend

* Changed KPI WAN (#4)

* Guardando cambios en la rama fix-wan-fw

* Ready for deploy in production, FW, WAN, Backend, Frontend

* Ready for deploy in production, FW, WAN, Backend, Frontend

* Ready for deploy in production, FW, WAN, Backend, Frontend

* Indicador WAN partido en 2

+++ Front/src/components/AdminPanel/AdminForms/Clients/DeleteClient.jsx
@@ -6,6 +6,7 @@ import "../form.css";
 export const DeleteClient = () => {
   const [ip, setIp] = useState("");
   const [mensaje, setMensaje] = useState("");
+  const token = localStorage.getItem("jwtToken");
 
   const handleIpChange = (event) => {
     setIp(event.target.value);
@@ -13,14 +14,23 @@ export const DeleteClient = () => {
 
   const handleSubmit = async (event) => {
     event.preventDefault();
-
+    // ! ###################
+    if (!token) {
+      setMensaje("No autorizado, por favor inicie sesión.");
+      return;
+    }
     try {
       if (ip.trim() === "") {
         setMensaje("Por favor, ingrese una dirección IP válida.");
         return; // No realizar la solicitud si ip está vacío
       }
       const response = await axios.delete(
-        `${BASE_API_URL}/clients/remove/${ip}`
+        `${BASE_API_URL}/clients/remove/${ip}`,
+        {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        }
       );
       setMensaje(response.data.message);
       setIp("");
